% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decaf_client.R
\name{build_http_header_user_agent}
\alias{build_http_header_user_agent}
\title{Builds and returns the "User-Agent" HTTP header for the HTTP calls
originating from this library.}
\usage{
build_http_header_user_agent()
}
\value{
A named-list for \code{User-Agent} key-value pair.
}
\description{
Builds and returns the "User-Agent" HTTP header for the HTTP calls
originating from this library.
}
