% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decaf_client.R
\name{build_http_header_x_decaf_url}
\alias{build_http_header_x_decaf_url}
\title{Builds and returns the "X-DECAF-URL" HTTP header for the HTTP calls
originating from this library.}
\usage{
build_http_header_x_decaf_url(url)
}
\arguments{
\item{url}{Base URL of the remote DECAF Instance.}
}
\value{
A named-list for \code{X-DECAF-URL}, \code{X-DECAF-APIURL} and
\code{X-DECAF-API-URL} key-value pairs.
}
\description{
For historical reasons, this function also adds \code{X-DECAF-APIURL} and
\code{X-DECAF-API-URL} headers, too. These will be removed at some point in the
future.
}
