% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{deleteResource}
\alias{deleteResource}
\title{Delete a resource to DECAF API.}
\usage{
deleteResource(..., params = list(), session = NULL)
}
\arguments{
\item{...}{URL path segments.}

\item{params}{Query parameters.}

\item{session}{Session information.}
}
\value{
Parsed R object for the response.
}
\description{
Delete a resource to DECAF API.
}
