% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\docType{data}
\name{.excludeTypes}
\alias{.excludeTypes}
\title{Defines a vector of types which we are not interesting in getting OHLC Observations for.}
\format{
An object of class \code{character} of length 9.
}
\usage{
.excludeTypes
}
\description{
Defines a vector of types which we are not interesting in getting OHLC Observations for.
}
\keyword{datasets}
