% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_utils.R
\name{.isNumeric}
\alias{.isNumeric}
\title{Indicates if the given argument is all numeric.}
\usage{
.isNumeric(x)
}
\arguments{
\item{x}{Argument to be checked.}
}
\value{
Logical indicating if the argument is numeric or not.
}
\description{
Indicates if the given argument is all numeric.
}
