% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.urlize}
\alias{.urlize}
\title{Make a URL.}
\usage{
.urlize(baseurl, segments, params = NULL)
}
\arguments{
\item{baseurl}{Base URL.}

\item{segments}{A vector of URL segments.}

\item{params}{Query string parameters.}
}
\value{
URL.
}
\description{
Make a URL.
}
