% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_utils.R
\name{get_header_value}
\alias{get_header_value}
\title{Finds and returns the value(s) of the given case-insensitive header name for
the given named-list of HTTP headers.}
\usage{
get_header_value(headers, name)
}
\arguments{
\item{headers}{A named-list of header name/value tuples.}

\item{name}{Header name.}
}
\value{
A vector of values.
}
\description{
Finds and returns the value(s) of the given case-insensitive header name for
the given named-list of HTTP headers.
}
