% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_queries.R
\name{getAccount}
\alias{getAccount}
\title{Returns the account identified by the given identifier or name.}
\usage{
getAccount(idOrName, session = NULL)
}
\arguments{
\item{idOrName}{Account ID or name}

\item{session}{Session which we will retrieve the remote resource within.}
}
\value{
The account of interest, if found.
}
\description{
Returns the account identified by the given identifier or name.
}
