% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getAccountByID}
\alias{getAccountByID}
\title{Returns the account identified by the given identifier.}
\usage{
getAccountByID(id, session = NULL)
}
\arguments{
\item{id}{Account ID}

\item{session}{Session which we will retrieve the remote resource within.}
}
\value{
The account of interest, if found.
}
\description{
Returns the account identified by the given identifier.
}
