% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getAccountByName}
\alias{getAccountByName}
\title{Returns the account identified by the given name.}
\usage{
getAccountByName(name, session = NULL)
}
\arguments{
\item{name}{Account name}

\item{session}{Session which we will retrieve the remote resource within.}
}
\value{
The account of interest, if found.
}
\description{
Returns the account identified by the given name.
}
