% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getArtifact}
\alias{getArtifact}
\title{Returns the artifact identified by the given identifier or symbol.}
\usage{
getArtifact(idOrSymbol, session = NULL)
}
\arguments{
\item{idOrSymbol}{Artifact ID or symbol}

\item{session}{Session which we will retrieve the remote resource within.}
}
\value{
The artifact of interest, if found.
}
\description{
Returns the artifact identified by the given identifier or symbol.
}
