% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getArtifactByID}
\alias{getArtifactByID}
\title{Returns the artifact identified by the given identifier.}
\usage{
getArtifactByID(id, session = NULL)
}
\arguments{
\item{id}{Artifact ID}

\item{session}{Session which we will retrieve the remote resource within.}
}
\value{
The artifact of interest, if found.
}
\description{
Returns the artifact identified by the given identifier.
}
