% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getArtifactBySymbol}
\alias{getArtifactBySymbol}
\title{Returns the artifact identified by the given symbol.}
\usage{
getArtifactBySymbol(symbol, session = NULL)
}
\arguments{
\item{symbol}{Artifact symbol}

\item{session}{Session which we will retrieve the remote resource within.}
}
\value{
The artifact of interest, if found.
}
\description{
Returns the artifact identified by the given symbol.
}
