% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_queries.R
\name{getLedger}
\alias{getLedger}
\title{Returns the ledger for the given query.}
\usage{
getLedger(
  artifact,
  accounts,
  start = NULL,
  end = NULL,
  type = "commitment",
  session = NULL
)
}
\arguments{
\item{artifact}{The artifact which we want the ledger for.}

\item{accounts}{Vector of account identifiers.}

\item{start}{Starting date of the ledger.}

\item{end}{Ending date of the ledger.}

\item{type}{Type of date of ledger.}

\item{session}{Session which we will retrieve the remote resource within.}
}
\value{
A ledger table.
}
\description{
Returns the ledger for the given query.
}
