% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{getResource}
\alias{getResource}
\title{Retrieve a resource from DECAF API.}
\usage{
getResource(..., params = list(), location = NULL, username = NULL,
  password = NULL, profile = NULL)
}
\arguments{
\item{...}{URL path segments.}

\item{params}{Query parameters.}

\item{location}{Base URL of the API.}

\item{username}{Username.}

\item{password}{Password.}

\item{profile}{Profile name.}
}
\value{
Parsed R object for the response.
}
\description{
Retrieve a resource from DECAF API.
}
