% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_prices.R
\name{getSymbolsOfInterest}
\alias{getSymbolsOfInterest}
\title{Returns a list of symbols of interest for retrieving OHLC
Observations for.}
\usage{
getSymbolsOfInterest(session = NULL)
}
\arguments{
\item{session}{Session data to communicate with DECAF server.}
}
\value{
A character vector of symbols.
}
\description{
Returns a list of symbols of interest for retrieving OHLC
Observations for.
}
