% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decaf_auth.R
\name{make_auth_header_value}
\alias{make_auth_header_value}
\title{Builds and returns 'Authorization' header value for the given credentials.}
\usage{
make_auth_header_value(x)
}
\arguments{
\item{x}{An instance of 'decaf.client.credentials' S3 class.}
}
\value{
A character string that can be used as 'Authorization' header value.
}
\description{
Builds and returns 'Authorization' header value for the given credentials.
}
