% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decaf_auth.R
\name{make_credentials}
\alias{make_credentials}
\title{Attempts to create DECAF API client credentials from given arguments.}
\usage{
make_credentials(
  username = NULL,
  password = NULL,
  apikey = NULL,
  apisecret = NULL,
  token = NULL,
  header = NULL
)
}
\arguments{
\item{username}{Username of "Basic Authentication" method.}

\item{password}{Password of "Basic Authentication" method.}

\item{apikey}{API key of "API Key" method.}

\item{apisecret}{API secret of "API Key" method.}

\item{token}{Token value of "API Token" method.}

\item{header}{Authorization header value to be used as is.}
}
\value{
An instance of 'decaf.client.credentials' S3 class with type
attribute set to one of \code{BASIC}, \code{APIKEY}, \code{TOKEN} or \code{HEADER}.
}
\description{
Attempts to create DECAF API client credentials from given arguments.
}
