% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decaf_auth.R
\name{make_credentials_inner}
\alias{make_credentials_inner}
\title{'decaf.client.credentials' S3 class constructor.}
\usage{
make_credentials_inner(x, type = c("BASIC", "APIKEY", "TOKEN", "HEADER"))
}
\arguments{
\item{x}{A named-list of credentials.}

\item{type}{Type of credentials.}
}
\value{
An instance of 'decaf.client.credentials' S3 class with type
attribute set to one of \code{BASIC}, \code{APIKEY}, \code{TOKEN} or \code{HEADER}.
}
\description{
'decaf.client.credentials' S3 class constructor.
}
