% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{makeSession}
\alias{makeSession}
\title{Attempts to create a new session with the provided arguments.}
\usage{
makeSession(location = NULL, token = NULL, username = NULL,
  password = NULL, profile = NULL, config = .defaultConfigFilepath,
  save = FALSE)
}
\arguments{
\item{location}{The location of the remote API base endpoint.}

\item{token}{Authentication token.}

\item{username}{Username.}

\item{password}{Password.}

\item{profile}{The profile name.}

\item{config}{The configuration file path.}

\item{save}{Indicates if we want to save the session.}
}
\value{
A session.
}
\description{
Attempts to create a new session with the provided arguments.
}
