% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_base.R
\name{patchResource}
\alias{patchResource}
\title{Partially update a resource to DECAF API.}
\usage{
patchResource(..., params = list(), payload = NULL, session = NULL)
}
\arguments{
\item{...}{URL path segments.}

\item{params}{Query parameters.}

\item{payload}{The payload to be send.}

\item{session}{Session information.}
}
\value{
Parsed R object for the response.
}
\description{
Partially update a resource to DECAF API.
}
