% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_base.R
\name{postMultipart}
\alias{postMultipart}
\title{Uploads a resource to DECAF API.}
\usage{
postMultipart(..., params = list(), payload = NULL, session = NULL)
}
\arguments{
\item{...}{URL path segments.}

\item{params}{Query parameters.}

\item{payload}{The payload to be send.}

\item{session}{Session information.}
}
\value{
Parsed R object for the response.
}
\description{
This is a variant of postResource method.
}
