% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_base.R
\name{postResource}
\alias{postResource}
\title{Send a resource to DECAF API.}
\usage{
postResource(..., params = list(), payload = NULL, session = NULL)
}
\arguments{
\item{...}{URL path segments.}

\item{params}{Query parameters.}

\item{payload}{The payload to be send.}

\item{session}{Session information.}
}
\value{
Parsed R object for the response.
}
\description{
Send a resource to DECAF API.
}
