% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decaf_auth.R
\name{print.decaf.client.credentials}
\alias{print.decaf.client.credentials}
\title{Pretty-prints authentication credentials by hiding the actual credentials.}
\usage{
\method{print}{decaf.client.credentials}(x)
}
\arguments{
\item{x}{An instance of 'decaf.client.credentials' S3 class.}
}
\description{
Pretty-prints authentication credentials by hiding the actual credentials.
}
