% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{putResource}
\alias{putResource}
\title{Update a resource to DECAF API.}
\usage{
putResource(..., params = list(), payload = NULL, session = NULL)
}
\arguments{
\item{...}{URL path segments.}

\item{params}{Query parameters.}

\item{payload}{The payload to be send.}

\item{session}{Session information.}
}
\value{
Parsed R object for the response.
}
\description{
Update a resource to DECAF API.
}
