% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{readConfig}
\alias{readConfig}
\title{Reads and returns the configuration file.}
\usage{
readConfig(filepath = .defaultConfigFilepath)
}
\arguments{
\item{filepath}{The path to the configuration file.}
}
\value{
Configuration if configuration file is read successfully,
    stops with error otherwise.
}
\description{
Reads and returns the configuration file.
}
