% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{readProfile}
\alias{readProfile}
\title{Reads a profile from the configuration file and returns it.}
\usage{
readProfile(name = "default", filepath = .defaultConfigFilepath)
}
\arguments{
\item{name}{The name of the profile.}

\item{filepath}{The path to the configuration file.}
}
\value{
The profile if configuration file is read and profile is
retrieved successfully, stops with error otherwise.
}
\description{
Reads a profile from the configuration file and returns it.
}
