% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{readProfiles}
\alias{readProfiles}
\title{Reads profiles from the configuration file and returns them.}
\usage{
readProfiles(filepath = .defaultConfigFilepath)
}
\arguments{
\item{filepath}{The path to the configuration file.}
}
\value{
List of profiles if configuration file is read successfully,
stops with error otherwise.
}
\description{
Reads profiles from the configuration file and returns them.
}
