% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_prices.R
\name{uploadOHLC}
\alias{uploadOHLC}
\title{Uploads a data frame of OHLC observations to the system.}
\usage{
uploadOHLC(observations, session = NULL)
}
\arguments{
\item{observations}{Observations.}

\item{session}{Session data to communicate with DECAF server.}
}
\value{
Result of uploading.
}
\description{
The \code{observations} argumens must be a data frame with
\code{date}, \code{symbol} and \code{close} columns. Additionally
and optionally it can include \code{open}, \code{high},
\code{low}, \code{volume} and \code{interest} columns.
}
