% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_base.R
\name{getResource}
\alias{getResource}
\title{Retrieve a resource from DECAF API.}
\usage{
getResource(..., params = list(), session = NULL)
}
\arguments{
\item{...}{URL path segments.}

\item{params}{Query parameters.}

\item{session}{Session information.}
}
\value{
Parsed R object for the response.
}
\description{
Retrieve a resource from DECAF API.
}
