% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr_session.R
\name{getSession}
\alias{getSession}
\title{Gets or creates a session.}
\usage{
getSession(
  location = NULL,
  apikey = NULL,
  apisecret = NULL,
  token = NULL,
  profile = "default",
  config = .defaultConfigFilepath,
  reset = FALSE
)
}
\arguments{
\item{location}{Location of the remote API base endpoint.}

\item{apikey}{API Key.}

\item{apisecret}{API Secret.}

\item{token}{Authentication token.}

\item{profile}{Profile name.}

\item{config}{Configuration file path.}

\item{reset}{Indicates if we are resetting the session first.}
}
\value{
Existing or new session.
}
\description{
Gets or creates a session.
}
