% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{getSession}
\alias{getSession}
\title{Gets or creates a session.}
\usage{
getSession(location = NULL, token = NULL, username = NULL,
  password = NULL, profile = "default", config = .defaultConfigFilepath,
  reset = FALSE)
}
\arguments{
\item{location}{The API Base URL.}

\item{token}{The API authentication token.}

\item{username}{The API authentication credential for username.}

\item{password}{The API authentication credential for password.}

\item{profile}{Profile name.}

\item{config}{The file path of the configuration.}

\item{reset}{Indicates if we are resetting the session first.}
}
\value{
Existing or new session.
}
\description{
Gets or creates a session.
}
