% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{makeSession}
\alias{makeSession}
\title{Attempts to create a new session with the provided arguments.}
\usage{
makeSession(location = NULL, apikey = NULL, apisecret = NULL,
  token = NULL, profile = NULL, config = .defaultConfigFilepath,
  save = FALSE)
}
\arguments{
\item{location}{Location of the remote API base endpoint.}

\item{apikey}{API Key.}

\item{apisecret}{API Secret.}

\item{token}{Authentication token.}

\item{profile}{Profile name.}

\item{config}{Configuration file path.}

\item{save}{Indicates if we want to save the session.}
}
\value{
A session.
}
\description{
Attempts to create a new session with the provided arguments.
}
